implements struts lab

+++ src/main/java/org/superbiz/struts/UserService.java
@@ -19,11 +19,12 @@ package org.superbiz.struts;
 
 import java.util.List;
 
+
 public interface UserService {
 
     public void add(User user);
 
-    public User find(int id);
+    public User find(long id);
 
     public List<User> findAll();
 }
